use resolver in provider
